requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.2
billiard==4.1.0
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.2
daphne==4.0.0
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-environ==0.10.0
django-timezone-field==5.1
djangorestframework==3.14.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.1
msgpack==1.0.5
nest-asyncio==1.5.6
prompt-toolkit==3.0.39
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.6.0
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
zope.interface==6.0
